requirements.txt
scikit-learn
tensorflow
wandb
numpy
pandas
Pillow
matplotlib
seaborn
opencv-python